Switched num_cmps to double, since int was too small for the larger inputs; Added a main function to input_generator - compile this by itself and then run it individually for generating the proper input files; Added a few shell scripts to easily chain together several runs of the program; Added some organization for the input/output files - the input files are in inputs/ and the output files are in outputs/(name of sort)/;

+++ sort.h
@@ -12,12 +12,12 @@
 
 class Sort {
 protected:
-   int num_cmps; // number of comparisons performed in sort function
+   double num_cmps; // number of comparisons performed in sort function
 public:
    virtual void sort(int A[], int size) = 0;    // main entry point
    bool testIfSorted(int A[], int size);        // returns false if not sorted
                                                 // true otherwise
-   int getNumCmps() { return num_cmps; }        // returns # of comparisons
+   double getNumCmps() { return num_cmps; }        // returns # of comparisons
 };
 
 class SelectionSort:public Sort {	// SelectionSort class
